+++ packages/derivatives-consumer/src/streams/Streams/TradeStream.ts
@@ -41,10 +41,14 @@ export class TradeStream {
     marketId,
     callback,
     executionSide,
+    onEndCallback = () => {},
+    onStatusCallback = (_status) => {},
   }: {
     marketId: string
     executionSide?: TradeExecutionSide
     callback: TradeStreamCallback
+    onEndCallback: () => void
+    onStatusCallback: (status: any) => void
   }) {
     const request = new StreamTradesRequest()
     request.setMarketId(marketId)
@@ -58,6 +62,8 @@ export class TradeStream {
     stream.on('data', (response: StreamTradesResponse) => {
       callback(transformer(response))
     })
+    stream.on('end', onEndCallback)
+    stream.on('status', onStatusCallback)
 
     return stream
   }
@@ -67,11 +73,15 @@ export class TradeStream {
     subaccountId,
     callback,
     executionSide,
+    onEndCallback = () => {},
+    onStatusCallback = (_status) => {},
   }: {
     marketId: string
     subaccountId: string
     executionSide?: TradeExecutionSide
     callback: TradeStreamCallback
+    onEndCallback: () => void
+    onStatusCallback: (status: any) => void
   }) {
     const request = new StreamTradesRequest()
     request.setMarketId(marketId)
@@ -86,6 +96,8 @@ export class TradeStream {
     stream.on('data', (response: StreamTradesResponse) => {
       callback(transformer(response))
     })
+    stream.on('end', onEndCallback)
+    stream.on('status', onStatusCallback)
 
     return stream
   }