+++ packages/spot-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.51](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.50...@injectivelabs/spot-consumer@0.0.51) (2021-07-14)
+
+
+### Features
+
+* added hooks 'end' and 'status' on the streams ([c7e8f26](https://github.com/InjectiveLabs/injective-ts/commit/c7e8f261bd0ee3b058a25dc1ef091c94158f99a7))
+
+
+
+
+
 ## [0.0.50](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/spot-consumer@0.0.49...@injectivelabs/spot-consumer@0.0.50) (2021-07-14)