+++ packages/subaccount-consumer/CHANGELOG.md
@@ -3,6 +3,17 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.0.25](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/subaccount-consumer@0.0.24...@injectivelabs/subaccount-consumer@0.0.25) (2021-07-14)
+
+
+### Features
+
+* added hooks 'end' and 'status' on the streams ([c7e8f26](https://github.com/InjectiveLabs/injective-ts/commit/c7e8f261bd0ee3b058a25dc1ef091c94158f99a7))
+
+
+
+
+
 ## [0.0.24](https://github.com/InjectiveLabs/injective-ts/compare/@injectivelabs/subaccount-consumer@0.0.23...@injectivelabs/subaccount-consumer@0.0.24) (2021-07-14)
 
 **Note:** Version bump only for package @injectivelabs/subaccount-consumer